+++ src/AppBundle/Repository/SafeguardingRepository.php
@@ -1,0 +1,13 @@
+<?php
+namespace AppBundle\Repository;
+
+/**
+ * SafeguardingRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class SafeguardingRepository extends DDBaseRepository
+{
+
+}
